move SimpleController::Base concerns into their own folder
